fix(journal): robust collector wiring (own enabled-gated config, no @ConditionalOnBean) + coverage
