+++ src/main/kotlin/ru/driics/aitrade/config/TradeJournalPersistenceConfig.kt
@@ -4,21 +4,16 @@ import com.zaxxer.hikari.HikariConfig
 import com.zaxxer.hikari.HikariDataSource
 import liquibase.integration.spring.SpringLiquibase
 import org.springframework.beans.factory.annotation.Value
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate
-import ru.driics.aitrade.application.journal.TradeCloseCollector
-import ru.driics.aitrade.domain.model.TradingMode
-import ru.driics.aitrade.domain.ports.ClosedPositionsPort
 import ru.driics.aitrade.domain.ports.TradeJournalPort
 import ru.driics.aitrade.domain.ports.TradeJournalQueryPort
 import ru.driics.aitrade.infra.persistence.JdbcTradeJournal
 import ru.driics.aitrade.infra.persistence.JdbcTradeJournalQuery
 import ru.driics.aitrade.infra.persistence.NoOpTradeJournal
 import ru.driics.aitrade.infra.persistence.NoOpTradeJournalQuery
-import java.time.Clock
 import javax.sql.DataSource
 
 /**
@@ -69,24 +64,6 @@ class TradeJournalPersistenceConfig {
     @Bean
     fun tradeJournalQuery(tradeJournalJdbcTemplate: NamedParameterJdbcTemplate): TradeJournalQueryPort =
         JdbcTradeJournalQuery(tradeJournalJdbcTemplate)
-
-    @Bean
-    @ConditionalOnBean(ClosedPositionsPort::class)
-    fun tradeCloseCollector(
-        closedPositions: ClosedPositionsPort,
-        tradeJournal: TradeJournalPort,
-        tradeJournalQuery: TradeJournalQueryPort,
-        clock: Clock,
-        tradingProperties: TradingProperties,
-        okxProperties: OkxProperties,
-    ): TradeCloseCollector = TradeCloseCollector(
-        source = closedPositions,
-        journal = tradeJournal,
-        query = tradeJournalQuery,
-        clock = clock,
-        mode = TradingMode.resolve(tradingProperties.demoMode, okxProperties.paper),
-        demo = tradingProperties.demoMode,
-    )
 }
 
 /**
